Refactored production code from tests
