+++ BlogAPI/Model/Post.cs
@@ -1,0 +1,8 @@
+﻿using System;
+namespace BlogAPI.Model
+{
+    public class Post
+    {
+        
+    }
+}
