+++ BlogAPIServiceTests/BlogAPIServiceTests.csproj
@@ -18,4 +18,7 @@
       <HintPath>..\..\..\..\..\usr\local\share\dotnet\sdk\NuGetFallbackFolder\microsoft.aspnetcore.mvc.core\2.2.0\lib\netstandard2.0\Microsoft.AspNetCore.Mvc.Core.dll</HintPath>
     </Reference>
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\BlogAPI\BlogAPI.csproj" />
+  </ItemGroup>
 </Project>
